# docs/swagger.yaml
openapi: 3.0.0
info:
  title: PlatIAgro Pipelines API
  version: "0.0.2"
  description: >
    These are the docs for PlatIAgro Pipelines API.
    The endpoints below are usually accessed by the PlatIAgro Web-UI.
  license:
    name: "Apache 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
servers:
  - url: http://localhost:8080
tags:
  - name: "Deployment"
    description: >
      Manage deployment notebooks.
  - name: "Training"
    description: >
      Manage training notebooks.
paths:
  /deployments:
    get:
      summary: "List all deployments."
      tags:
        - "Deployment"
      responses:
        "200":
          $ref: "#/components/responses/Deployments"
        "500":
          $ref: "#/components/responses/InternalServerError"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
  /deployment:
    post:
      summary: "Run a deployment pipeline"
      tags:
        - "Deployment"
      requestBody:
        $ref: "#/components/requestBodies/DeploymentPost"
      responses:
        "200":
          $ref: "#/components/responses/Run"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalServerError"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
  /deployments/logs:
    get:
      summary: "Get logs from deployment."
      tags:
        - "Deployment"
      parameters:
        - name: experimentId
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          $ref: "#/components/responses/DeploymentLog"
        "500":
          $ref: "#/components/responses/InternalServerError"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
  /training/<run_id>:
    get:
      summary: "Get the training run status."
      tags:
        - "Training"
      parameters:
        - name: run_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          $ref: "#/components/responses/TrainingStatus"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalServerError"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
  /training:
    post:
      summary: "Run a training pipeline"
      tags:
        - "Training"
      requestBody:
        $ref: "#/components/requestBodies/TrainingPost"
      responses:
        "200":
          $ref: "#/components/responses/Run"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalServerError"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
components:
  schemas:
    Run:
      type: object
      properties:
        message:
          type: string
          example: "Pipeline running."
        runId:
          type: string
          format: uuid
    Deployments:
      type: array
      items:
        $ref: "#/components/schemas/DeploymentStatus"
    DeploymentStatus:
      type: object
      properties:
        experimentId:
          type: string
          format: uuid
        name:
          type: string
        status:
          type: string
        url:
          type: string
        createdAt:
          type: string
          format: date-time
    DeploymentLog:
      type: array
      items:
        type: object
        properties:
          containerName:
            type: string
          logs:
            type: array
            items:
              oneOf:
                - $ref: "#/components/schemas/Log"
    Log:
      type: object
      properties:
        level:
          type: string
        message:
          type: string
        timestamp:
          type: string
    TrainingStatus:
      type: object
      properties:
        status:
          type: object
          additionalProperties:
            type: string
          example:
            c9a4e6a05-a788-4efa-b94b-693e28608618: "Succeeded"
            caadb1d13-0666-4d84-a365-356b258e4433: "Succeeded"
  requestBodies:
    DeploymentPost:
      content:
        application/json:
          schema:
            type: object
            properties:
              experiment_id:
                type: string
              components:
                type: array
                items:
                  type: string
              dataset:
                type: string
              target:
                type: string
    TrainingPost:
      content:
        application/json:
          schema:
            type: object
            properties:
              experiment_id:
                type: string
              components:
                type: array
                items:
                  type: string
              dataset:
                type: string
              target:
                type: string
  responses:
    DeploymentLog:
      description: ""
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/DeploymentLog"
    Run:
      description: ""
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Run"
    Deployments:
      description: ""
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Deployments"
    TrainingStatus:
      description: ""
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/TrainingStatus"
    Message:
      description: ""
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
            required:
              - message
    BadRequest:
      description: ""
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
            required:
              - message
    NotFound:
      description: ""
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                example: "The specified ... does not exist"
            required:
              - message
    InternalServerError:
      description: ""
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                example: "An internal failure occurred."
            required:
              - message
    ServiceUnavailable:
      description: ""
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                example: "The service is unavailable. Try your call again."
            required:
              - message